Small tweaks with no noticeable behavioral changes

+++ include/reni/pal/win32/Win32Window.hpp
@@ -119,7 +119,7 @@ namespace reni::pal::win32 {
 			POINT p = {};
 			safeApiCall(GetCursorPos(&p));
 			safeApiCall(ScreenToClient(m_handle.get(), &p));
-			return { static_cast<float>(p.x), static_cast<float>(p.y) };
+			return { p.x, p.y };
 		}
 
 
@@ -159,19 +159,30 @@ namespace reni::pal::win32 {
 
 					// mouse messages
 					case WM_MOUSEMOVE:
-						return window->performCallback(&WindowCallbacks::onMouseMove, Point2(
-							static_cast<float>(GET_X_LPARAM(lParam)),
-							static_cast<float>(GET_Y_LPARAM(lParam))
-						));
+						return window->performCallback(&WindowCallbacks::onMouseMove,
+							Point2(
+								GET_X_LPARAM(lParam),
+								GET_Y_LPARAM(lParam)
+							)
+						);
 
 					case WM_LBUTTONDOWN: case WM_LBUTTONUP:
-						return window->performCallback(&WindowCallbacks::onMouseButtonStateChange, MouseButtons::Left, msg == WM_LBUTTONDOWN);
+						return window->performCallback(&WindowCallbacks::onMouseButtonStateChange,
+							MouseButtons::Left,
+							msg == WM_LBUTTONDOWN
+						);
 
 					case WM_RBUTTONDOWN: case WM_RBUTTONUP:
-						return window->performCallback(&WindowCallbacks::onMouseButtonStateChange, MouseButtons::Right, msg == WM_RBUTTONDOWN);
+						return window->performCallback(&WindowCallbacks::onMouseButtonStateChange,
+							MouseButtons::Right,
+							msg == WM_RBUTTONDOWN
+						);
 
 					case WM_MBUTTONDOWN: case WM_MBUTTONUP:
-						return window->performCallback(&WindowCallbacks::onMouseButtonStateChange, MouseButtons::Middle, msg == WM_MBUTTONDOWN);
+						return window->performCallback(&WindowCallbacks::onMouseButtonStateChange,
+							MouseButtons::Middle,
+							msg == WM_MBUTTONDOWN
+						);
 				}
 			}
 			return DefWindowProc(hwnd, msg, wParam, lParam);
